+++ tapis-appsapi/src/main/java/edu/utexas/tacc/tapis/apps/api/resources/AppResource.java
@@ -1108,6 +1108,7 @@ public class AppResource
           jobAttrs.description, jobAttrs.dynamicExecSystem, jobAttrs.execSystemConstraints, jobAttrs.execSystemId,
           jobAttrs.execSystemExecDir, jobAttrs.execSystemInputDir, jobAttrs.execSystemOutputDir,
           jobAttrs.execSystemLogicalQueue, jobAttrs.archiveSystemId, jobAttrs.archiveSystemDir, jobAttrs.archiveOnAppError,
+          jobAttrs.isMpi, jobAttrs.mpiCmd, jobAttrs.cmdPrefix,
           jobAttrs.parameterSet, jobAttrs.fileInputs, jobAttrs.fileInputArrays, jobAttrs.nodeCount, jobAttrs.coresPerNode,
           jobAttrs.memoryMB, jobAttrs.maxMinutes, jobAttrs.subscriptions, jobAttrs.tags,
           req.tags, notes, null, false, null, null);
@@ -1133,6 +1134,7 @@ public class AppResource
           jobAttrs.description, jobAttrs.dynamicExecSystem, jobAttrs.execSystemConstraints, jobAttrs.execSystemId,
           jobAttrs.execSystemExecDir, jobAttrs.execSystemInputDir, jobAttrs.execSystemOutputDir,
           jobAttrs.execSystemLogicalQueue, jobAttrs.archiveSystemId, jobAttrs.archiveSystemDir, jobAttrs.archiveOnAppError,
+          jobAttrs.isMpi, jobAttrs.mpiCmd, jobAttrs.cmdPrefix,
           jobAttrs.parameterSet, jobAttrs.fileInputs, jobAttrs.fileInputArrays, jobAttrs.nodeCount, jobAttrs.coresPerNode,
           jobAttrs.memoryMB, jobAttrs.maxMinutes, jobAttrs.subscriptions, jobAttrs.tags,
           req.tags, notes, null, false, null, null);