+++ tapis-appslib/src/main/java/edu/utexas/tacc/tapis/apps/dao/AppsDaoImpl.java
@@ -212,6 +212,9 @@ public class AppsDaoImpl extends AbstractDao implements AppsDao
               .set(APPS_VERSIONS.ARCHIVE_SYSTEM_ID, app.getArchiveSystemId())
               .set(APPS_VERSIONS.ARCHIVE_SYSTEM_DIR, app.getArchiveSystemDir())
               .set(APPS_VERSIONS.ARCHIVE_ON_APP_ERROR, app.isArchiveOnAppError())
+              .set(APPS_VERSIONS.IS_MPI, app.getIsMpi())
+              .set(APPS_VERSIONS.MPI_CMD, app.getMpiCmd())
+              .set(APPS_VERSIONS.CMD_PREFIX, app.getCmdPrefix())
               .set(APPS_VERSIONS.PARAMETER_SET, parameterSetJson)
               .set(APPS_VERSIONS.FILE_INPUTS, fileInputsJson)
               .set(APPS_VERSIONS.FILE_INPUT_ARRAYS, fileInputArraysJson)
@@ -347,6 +350,9 @@ public class AppsDaoImpl extends AbstractDao implements AppsDao
               .set(APPS_VERSIONS.ARCHIVE_SYSTEM_ID, putApp.getArchiveSystemId())
               .set(APPS_VERSIONS.ARCHIVE_SYSTEM_DIR, putApp.getArchiveSystemDir())
               .set(APPS_VERSIONS.ARCHIVE_ON_APP_ERROR, putApp.isArchiveOnAppError())
+              .set(APPS_VERSIONS.IS_MPI, putApp.getIsMpi())
+              .set(APPS_VERSIONS.MPI_CMD, putApp.getMpiCmd())
+              .set(APPS_VERSIONS.CMD_PREFIX, putApp.getCmdPrefix())
               .set(APPS_VERSIONS.PARAMETER_SET, parameterSetJson)
               .set(APPS_VERSIONS.FILE_INPUTS, fileInputsJson)
               .set(APPS_VERSIONS.FILE_INPUT_ARRAYS, fileInputArraysJson)
@@ -474,6 +480,9 @@ public class AppsDaoImpl extends AbstractDao implements AppsDao
               .set(APPS_VERSIONS.ARCHIVE_SYSTEM_ID, patchedApp.getArchiveSystemId())
               .set(APPS_VERSIONS.ARCHIVE_SYSTEM_DIR, patchedApp.getArchiveSystemDir())
               .set(APPS_VERSIONS.ARCHIVE_ON_APP_ERROR, patchedApp.isArchiveOnAppError())
+              .set(APPS_VERSIONS.IS_MPI, patchedApp.getIsMpi())
+              .set(APPS_VERSIONS.MPI_CMD, patchedApp.getMpiCmd())
+              .set(APPS_VERSIONS.CMD_PREFIX, patchedApp.getCmdPrefix())
               .set(APPS_VERSIONS.PARAMETER_SET, parameterSetJson)
               .set(APPS_VERSIONS.FILE_INPUTS, fileInputsJson)
               .set(APPS_VERSIONS.FILE_INPUT_ARRAYS, fileInputArraysJson)
@@ -1561,6 +1570,7 @@ public class AppsDaoImpl extends AbstractDao implements AppsDao
             r.get(APPS_VERSIONS.EXEC_SYSTEM_INPUT_DIR), r.get(APPS_VERSIONS.EXEC_SYSTEM_OUTPUT_DIR),
             r.get(APPS_VERSIONS.EXEC_SYSTEM_LOGICAL_QUEUE), r.get(APPS_VERSIONS.ARCHIVE_SYSTEM_ID),
             r.get(APPS_VERSIONS.ARCHIVE_SYSTEM_DIR), r.get(APPS_VERSIONS.ARCHIVE_ON_APP_ERROR),
+            r.get(APPS_VERSIONS.IS_MPI), r.get(APPS_VERSIONS.MPI_CMD), r.get(APPS_VERSIONS.CMD_PREFIX),
             parmSet, fileInputs, fileInputArrays, r.get(APPS_VERSIONS.NODE_COUNT), r.get(APPS_VERSIONS.CORES_PER_NODE),
             r.get(APPS_VERSIONS.MEMORY_MB), r.get(APPS_VERSIONS.MAX_MINUTES), subscriptions,
             r.get(APPS_VERSIONS.JOB_TAGS), r.get(APPS_VERSIONS.TAGS), r.get(APPS_VERSIONS.NOTES),